requirements.txt
Flask
python-dotenv
requests
bs4
random